#11793 Delete User Dependency before delete the User https://github.com/metasfresh/metasfresh/issues/11793 - prevent delete logged-in user

+++ backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/adempiere/modelvalidator/AD_User.java
@@ -2,13 +2,17 @@ package de.metas.adempiere.modelvalidator;
 
 import de.metas.bpartner.BPartnerId;
 import de.metas.bpartner.service.IBPartnerBL;
+import de.metas.i18n.AdMessageKey;
+import de.metas.i18n.IMsgBL;
+import de.metas.i18n.ITranslatableString;
 import de.metas.i18n.Language;
+import de.metas.i18n.TranslatableStrings;
 import de.metas.title.Title;
 import de.metas.title.TitleId;
 import de.metas.title.TitleRepository;
+import de.metas.user.UserId;
 import de.metas.user.UserPOCopyRecordSupport;
 import de.metas.user.api.IUserBL;
-import de.metas.user.api.IUserDAO;
 import de.metas.util.Check;
 import de.metas.util.Services;
 import lombok.NonNull;
@@ -18,10 +22,12 @@ import org.adempiere.ad.callout.spi.IProgramaticCalloutProvider;
 import org.adempiere.ad.modelvalidator.annotations.Init;
 import org.adempiere.ad.modelvalidator.annotations.Interceptor;
 import org.adempiere.ad.modelvalidator.annotations.ModelChange;
+import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.model.CopyRecordFactory;
 import org.compiere.SpringContextHolder;
 import org.compiere.model.I_AD_User;
 import org.compiere.model.ModelValidator;
+import org.compiere.util.Env;
 
 import java.util.Optional;
 
@@ -36,10 +42,13 @@ import java.util.Optional;
 @Callout(I_AD_User.class)
 public class AD_User
 {
+	private static final AdMessageKey MSG_UserDelete = AdMessageKey.of("UserDeleteMsg");
+
 	private final IBPartnerBL bpPartnerService = Services.get(IBPartnerBL.class);
 	private final TitleRepository titleRepository = SpringContextHolder.instance.getBean(TitleRepository.class);
-	private final IUserDAO userDAO = Services.get(IUserDAO.class);
 	private final IUserBL userBL = Services.get(IUserBL.class);
+	private final IMsgBL msgBL = Services.get(IMsgBL.class);
+
 	@Init
 	public void init()
 	{
@@ -109,7 +118,23 @@ public class AD_User
 	@ModelChange(timings = {ModelValidator.TYPE_BEFORE_DELETE})
 	public void beforeDelete(@NonNull final I_AD_User userRecord)
 	{
-		userBL.deleteUserDependency(userRecord);
+		UserId loggedInUserId = Env.getLoggedUserIdIfExists().orElse(null);
+		if (loggedInUserId != null )
+		{
+			if (loggedInUserId.getRepoId() == userRecord.getAD_User_ID())
+			{
+				final ITranslatableString errorMsg = msgBL.getTranslatableMsgText(MSG_UserDelete);
+				throw new AdempiereException(TranslatableStrings.builder()
+													 .append(errorMsg)
+													 .build())
+						.appendParametersToMessage()
+						.setParameter("AD_User_ID", userRecord.getAD_User_ID())
+						.setParameter("Name", userRecord.getName());
+			}
+
+			userBL.deleteUserDependency(userRecord);
+		}
+
 	}
 
 	@ModelChange(timings = { ModelValidator.TYPE_AFTER_DELETE },
